+++ src/main/java/br/com/soapboxrace/engine/Personas.java
@@ -20,8 +20,8 @@ public class Personas extends Router {
 	private long getPersonaId(boolean isBypass) throws PersonaIdMismatchException {
 		String[] targetSplitted = getTarget().split("/");
 		Long idPersona = Long.valueOf(targetSplitted[4]);
-		if (((isBypass || idPersona.equals(getLoggedPersonaId()) || getLoggedPersonaId() == (long) -1)))
-			if (getUserId() != (long) -1 && getSecurityToken() != null
+		if (((isBypass || idPersona.equals(getLoggedPersonaId()) || getLoggedPersonaId() == -1L)))
+			if (getUserId() != -1L && getSecurityToken() != null
 					&& Router.activeUsers.get(getUserId()).getSecurityToken().equals(getSecurityToken()))
 				return idPersona;
 		throw new ServerExceptions.PersonaIdMismatchException(getLoggedPersonaId(), idPersona);
